src/requirements.txt
fastapi
uvicorn[standard]
litellm
langgraph
langchain-core
pydantic
python-dotenv
supabase
httpx
structlog
